+++ temp/useTenantKy/useTenantKy.js
@@ -1,0 +1,19 @@
+import { useOkapiKy } from '@folio/stripes/core';
+
+const useTenantKy = ({ tenantId } = {}) => {
+  const ky = useOkapiKy();
+
+  return tenantId
+    ? ky.extend({
+      hooks: {
+        beforeRequest: [
+          request => {
+            request.headers.set('X-Okapi-Tenant', tenantId);
+          },
+        ],
+      },
+    })
+    : ky;
+};
+
+export default useTenantKy;